+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.1.0 (Unreleased)
+
+* Feature: ingest log files from a container
+
+* Helm: add new property `kubeletRoot`
+
 ## 1.0.1
 
 * install plugin `fluent-plugin-concat` in Fluentd